Module Mini_graph

Mini graph module

type graph = int list array

Abstract type of graphs

val iter : (int -> int list -> unit) -> graph -> unit

Iterate over vertices and their successors list

val succ : graph -> int -> int list

Get the sucessors of a given vertex

val mk_graph : (int * int) list -> graph

Create a new graph based on a list of (oriented) edges

val vertices : graph -> int

Return the number of vertices in a graph (in time O(1))

val dfs : graph -> (int -> unit) -> (int -> unit) -> unit

Perform a Depth-first-search traversal. dfs g pre post will execute pre x when x is first discovered and post x right after the last successor of x has been visited

val dfs_ordered : graph -> (int -> unit) -> (int -> unit) -> int list -> unit

Perform a Depth-first-search traversal by following a given order

val topo_sort : graph -> int list

Perform topoligical sort on a graph. Returns an ordered list of vertices

exception Cycle

Exception raised when a cycle is detected in a graph

val has_cycle : graph -> bool

Detect cycles in a graph

val transpose : graph -> int list array

Compute the transposed graph of a graph

val kosaraju : graph -> int list list

Computes the strongly connected components of a graph

val bfs : graph -> int -> (int -> unit) -> unit

Perform a Breadth-first-search traversal

val dijkstra : graph -> int -> ((int * int) -> float) -> float array * int option array

dikstra g s cost returns a tuple distances, preds where distances.(x) is the cost of an optimal path from s to x in g and preds.(x) is the predecessor of x in the optimal path from s to x.

module Clique : Stdlib.Set.S with type Clique.elt = int

A clique in a graph is a set of vertices

module CliqueSet : Stdlib.Set.S with type CliqueSet.elt = Clique.t

Set of cliques

val bron_kerbosch : graph -> CliqueSet.t

Computes the set of all maximal cliques in a graph using Bron-Kerbosch algorithm.